requirements.txt
fastapi
uvicorn
scikit-learn
xgboost
pandas
numpy
networkx
joblib
python-multipart
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
websockets==12.0
plotly==5.18.0
dash==2.14.2 